ai-engine-python/pipeline/ingest_stats.py
import os
import psycopg2
from dotenv import load_dotenv

load_dotenv()

def get_db_connection():
    return psycopg2.connect(
        host=os.getenv('DB_HOST', 'localhost'),
        port=os.getenv('DB_PORT', '5432'),
        database=os.getenv('DB_NAME', 'postgres'),
        user=os.getenv('DB_USER', 'postgres'),
        password=os.getenv('DB_PASSWORD', 'postgres')
    )

def ingest_stats():
    print("Starting stats ingestion...")
    # Scrape stats
    print("Stats ingestion complete.")

if __name__ == '__main__':
    ingest_stats()
